refactor: Update header height and sidebar width
Adjust the header height from 94px to 40px in the _variables.custom.scss file. Also, update the sidebar width from 265px to 250px in the same file. These changes were made to improve the layout and make it more compact.

Refactor the sidebar toggle button styling

In the _sidebar.scss file, modify the box-shadow and background-color properties of the .app-sidebar-toggle class to improve the styling of the sidebar toggle button. This change affects the desktop mode layout.

Update the app-sidebar-logo height

In the HeaderWrapper.tsx file, change the height of the app-sidebar-logo class from 94px to 40px. This adjustment ensures consistency with the updated header height.

Remove unused code in Navbar.tsx

In the Navbar.tsx file, remove the import statement for the HeaderNotificationsMenu component, as it is no longer used in the codebase.